# read from socket to the buffer instead of a char