~ % benchstat old.txt new.txt
name             old time/op    new time/op    delta
_put_delete_1k      318µs ± 3%     260µs ± 3%  -18.38%  (p=0.000 n=9+8)
_put_delete_64k     574µs ± 2%     514µs ± 8%  -10.49%  (p=0.000 n=10+10)
_put_delete_8       306µs ± 1%     256µs ± 2%  -16.51%  (p=0.000 n=8+9)
_put_delete_8k      365µs ± 3%     306µs ± 5%  -16.10%  (p=0.000 n=9+10)

name             old speed      new speed      delta
_put_delete_1k   4.83MB/s ± 0%  4.97MB/s ± 1%   +2.94%  (p=0.000 n=9+9)
_put_delete_64k   129MB/s ± 0%   148MB/s ±28%  +14.36%  (p=0.000 n=10+10)
_put_delete_8    40.0kB/s ± 0%  40.0kB/s ± 0%     ~     (all equal)
_put_delete_8k   28.4MB/s ±34%  38.9MB/s ± 1%  +36.76%  (p=0.000 n=10+10)

Updates #430