@kr 2019-08-01
> It has improved the throughput just slightly...

Then I wonder if this change is worth the extra code… 🤔