@ysmolsky 2019-08-01
> > It has improved the throughput just slightly...
> 
> Then I wonder if this change is worth the extra code… 🤔

Same here. I have a suspicion that this code is not the slowest part of the benchmark (yet). So it's bottlenecked by something else. Also it might be that read is actually uses some kind of buffering already. On powerful Linux this code speeds up benchmarks up to 10%