@kr 2019-08-01
My understanding is that the kernel has network buffers for TCP connections and that, if those buffers don’t fill up, the throughout inprovement in this change comes from making fewer kernel calls, so fewer context switches.

But 10% is nothing to sneeze at.