@kr 2019-08-02
That is interesting! But those *p* values suggest the results might not be super reliable haha. It might need a higher repetition count?