@ysmolsky 2019-08-02
I thought that low or next to zero values of `p` are good, no? 
I have not ran it many times, n=10 there. 

I tried to run the same benchmark with n=50 and what I got. There are peaks of slowness in 64k benchmarks 
```
Benchmarkput_delete_64k     3000            490953 ns/op         187.73 MB/s
Benchmarkput_delete_64k     3000            488790 ns/op         187.84 MB/s
Benchmarkput_delete_64k     3000            462030 ns/op         189.30 MB/s
Benchmarkput_delete_64k     3000           1764607 ns/op          39.09 MB/s
Benchmarkput_delete_64k     3000            629254 ns/op         180.57 MB/s
Benchmarkput_delete_64k     3000            532570 ns/op         185.52 MB/s
Benchmarkput_delete_64k     1000           1190625 ns/op          64.31 MB/s
Benchmarkput_delete_64k     3000            723916 ns/op          97.47 MB/s
Benchmarkput_delete_64k     3000            550059 ns/op         184.60 MB/s
Benchmarkput_delete_64k     3000            741290 ns/op          97.21 MB/s
Benchmarkput_delete_64k     3000            529147 ns/op         185.70 MB/s
Benchmarkput_delete_64k     3000            535995 ns/op         185.34 MB/s
```
and multiple errors in 8k size:
```
start server port=51635 pid=48825
>1025 put 0 0 0 8192
>1025 a ...SKIP... a>1025
<1025 select: Invalid argument
error (exit status 1)
Benchmarkput_delete_8k
```
It happens both in master and this branch. I guess the high value of FD might be related, maybe FDs are not closed soon enough.