@ysmolsky 2019-08-02
I have found a workaround for this bug and run cttest multiple times from the shell:
```
make ct/_ctcheck ; repeat 25 { ct/_ctcheck -test . -bench 'ctbench_put_*' } > new.txt```
Results for n=25:
```
name            old time/op    new time/op    delta
put_delete_1k      314µs ± 3%     260µs ± 5%  -17.23%  (p=0.000 n=24+23)
put_delete_64k     530µs ± 3%     478µs ± 4%   -9.69%  (p=0.000 n=21+23)
put_delete_8       311µs ± 3%     255µs ± 2%  -17.96%  (p=0.000 n=24+24)
put_delete_8k      360µs ± 2%     307µs ± 3%  -14.62%  (p=0.000 n=25+24)

name            old speed      new speed      delta
put_delete_1k   4.84MB/s ± 0%  4.97MB/s ± 1%   +2.59%  (p=0.000 n=23+23)
put_delete_64k   186MB/s ± 0%   188MB/s ± 1%   +1.48%  (p=0.000 n=21+23)
put_delete_8    40.0kB/s ± 0%  40.0kB/s ± 0%     ~     (all equal)
put_delete_8k   37.9MB/s ± 0%  38.9MB/s ± 0%   +2.48%  (p=0.000 n=24+24)
```
Even though time/op has dropped > 10%, the speed has increased  < 3%