@kr 2019-08-02
> I thought that low or next to zero values of p are good, no?

Ah yes, of course you are right. Sorry about that!